+++ CHANGELOG.md
@@ -5,7 +5,7 @@ See [Keep a CHANGELOG](http://keepachangelog.com/) for more details.
 
 ## [Unreleased]
 ### Added
-- use fuzzywuzzy to match input strings
+- use fuzzywuzzy (if available) to match input strings
 ### Fixed
 - follow symlinks in the `password-filter.py` file [#18]
 - set transient Alfred clipboard [#14]